Utilize eloquents polymorphic relations
This change replaces five associative/pivot tables, with one groupables table. This will simplify the database structure and make the erd easier to read.

The downside is that we will have to rely on eloquent to maintain our database consistency, as I can no longer define foreign key constraints in the database itself. I think we can all agree that eloquent is trustworthy, however.