Prevented adding groupable to group > once
Utilizes the primary key constraint on the table and the resultant SQL exception if a duplicate record is attempted. I was unable to find an elegant way of checking through eloquent alone, due to the polymorphic nature of these relationships.

This checking is not required for detach() as this will naturally have no effect if the record doesn't exist.